Fixed issues related to correct functionality of the create form (DynamicForm)
Nested objects can now be created from the create form of a parent object, and automatically be selected from  the dropdown. Submitting a filled out form is now supported. There are still issues of lacking fields being sent in the API call. The objectConfig should be updated with all fields required for the  API calls.